#24 Move LogIn folder to page folder from components folder
